requirements.txt
alembic==1.5.8
apache-airflow==2.0.1
apache-airflow-providers-amazon==1.2.0
apache-airflow-providers-ftp==1.0.1
apache-airflow-providers-http==1.1.1
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-sqlite==1.0.2
apispec==3.3.2
argcomplete==1.12.2
attrs==20.3.0
Babel==2.9.0
boto3==1.15.18
botocore==1.18.18
cached-property==1.5.2
cattrs==1.4.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.7.0
croniter==0.3.37
cryptography==3.4.7
defusedxml==0.7.1
dill==0.3.3
dnspython==2.1.0
docutils==0.17
email-validator==1.1.2
Flask==1.1.2
Flask-AppBuilder==3.1.1
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.16
greenlet==1.0.0
gunicorn==19.10.0
idna==2.10
importlib-resources==1.5.0
inflection==0.5.1
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
natsort==7.1.1
numpy==1.20.2
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.0
pandas==1.2.3
pendulum==2.1.2
prison==0.1.3
psutil==5.8.0
pycparser==2.20
Pygments==2.8.1
PyJWT==1.7.1
pyrsistent==0.17.3
python-daemon==2.3.0
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.25.1
rich==9.2.0
s3transfer==0.3.6
setproctitle==1.2.2
six==1.15.0
SQLAlchemy==1.4.6
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.36.8
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.25.11
watchtower==0.7.3
Werkzeug==1.0.1
WTForms==2.3.3
